+++ services/blockchain-indexer/shared/constants.js
@@ -110,11 +110,6 @@ const KV_STORE_KEY = Object.freeze({
 });
 
 // TODO: Update values from sdk once this is settled: https://github.com/LiskHQ/discussions/discussions/101
-const CHAIN_STATUS = Object.freeze({
-	0: 'registered',
-	1: 'active',
-	2: 'terminated',
-});
 
 const TRANSACTION_STATUS = Object.freeze({
 	SUCCESS: 'success',
@@ -140,6 +135,5 @@ module.exports = {
 	COMMAND,
 	MAX_COMMISSION,
 	KV_STORE_KEY,
-	CHAIN_STATUS,
 	TRANSACTION_STATUS,
 };
