+++ services/gateway/apis/http-version3/swagger/definitions/blockchainApps.json
@@ -20,10 +20,10 @@
 				"example": "00000000",
 				"description": "Chain ID of the blockchain application."
 			},
-			"state": {
+			"status": {
 				"type": "string",
 				"example": "active",
-				"description": "Current state of the application"
+				"description": "Current status of the application"
 			},
 			"address": {
 				"type": "string",
